refactor(payment): 내부 통신 컨트롤러 명칭 변경 및 환경변수 설정 최적화
