+++ src/main/resources/application-dev.yml
@@ -28,7 +28,3 @@ logging:
     org.hibernate.orm.jdbc.bind: TRACE
     org.hibernate.orm.jdbc.extract: TRACE
     org.springframework.transaction.interceptor: TRACE
-
-custom:
-  global:
-    internalBackUrl: http://localhost:8080
