+++ src/main/resources/application.yml
@@ -11,3 +11,7 @@ spring:
   output:
     ansi:
       enabled: always
+
+custom:
+  global:
+    internalBackUrl: ${INTERNAL_BACK_URL:http://localhost:8080}
